make sure daemon exists while terminating

+++ run-amazfish.sh
@@ -3,6 +3,13 @@
 cd /app/bin
 ./harbour-amazfishd &
 daemon_pid=$!
-./harbour-amazfish-ui
 
-kill "$daemon_pid"
+function cleanup() {
+    if [[ -n "$daemon_pid" && -e "/proc/$daemon_pid/exe" ]]; then
+        echo "Terminating daemon $daemon_pid"
+        kill "$daemon_pid"
+    fi
+}
+trap cleanup EXIT
+
+./harbour-amazfish-ui
